aider: Remove standard library packages from `requirements.txt`.
# Aider chat conversation:

USER: Make sure `requirements.txt` doesn't contain packages that come with python like json os and sys

ASSISTANT: Here is the diff to update the `requirements.txt` file to remove the standard library packages:

```diff
--- requirements.txt
+++ requirements.txt
@@ ... @@
-pre-commitjson
-os
-sys
-threading
-queue
-subprocess
```

This diff removes all the lines from the `requirements.txt` file since all the listed packages are part of the Python standard library and do not need to be included in this file. If there are any third-party libraries that the daemon depends on, please provide their names, and I will add them to the `requirements.txt`.